change precision type of cosine_similarity to float64

+++ pkg/vectorize/moarray/array.go
@@ -195,7 +195,7 @@ func L2Norm[T types.RealNumbers](v []T) (float64, error) {
 	return math.Sqrt(float64(sum)), nil
 }
 
-func CosineSimilarity[T types.RealNumbers](v1, v2 []T) (float32, error) {
+func CosineSimilarity[T types.RealNumbers](v1, v2 []T) (float64, error) {
 
 	if len(v1) != len(v2) {
 		return 0, moerr.NewArrayInvalidOpNoCtx(len(v1), len(v2))
@@ -214,5 +214,5 @@ func CosineSimilarity[T types.RealNumbers](v1, v2 []T) (float32, error) {
 
 	// using math.Sqrt instead of momath.Sqrt() because argument of Sqrt will never be negative for real numbers.
 	// casting it to float32, because cosine_similarity is between 1 and -1.
-	return float32(float64(innerProduct) / math.Sqrt(float64(normV1*normV2))), nil
+	return float64(innerProduct) / math.Sqrt(float64(normV1*normV2)), nil
 }
